+++ lib/pages/settings/outbox/outbox_monitor.dart
@@ -122,11 +122,11 @@ class OutboxItemCard extends StatelessWidget {
     Color cardColor(OutboxStatus status) {
       switch (statusEnum) {
         case OutboxStatus.pending:
-          return styleConfig().outboxPendingColor;
+          return styleConfig().primaryColorLight;
         case OutboxStatus.error:
           return styleConfig().alarm;
         case OutboxStatus.sent:
-          return styleConfig().outboxSuccessColor;
+          return styleConfig().primaryColor;
       }
     }
 
@@ -137,6 +137,7 @@ class OutboxItemCard extends StatelessWidget {
     return Padding(
       padding: const EdgeInsets.all(2),
       child: Card(
+        color: cardColor(statusEnum),
         child: ListTile(
           contentPadding: const EdgeInsets.only(left: 24, right: 24),
           title: Text(